requirements.txt
aio-pika==6.4.1
aiohttp==3.6.2
aiormq==3.2.0
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
chardet==3.0.4
Click==7.0
coverage==5.0.1
entrypoints==0.3
flake8==3.7.9
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
multidict==4.7.3
mypy==0.761
mypy-extensions==0.4.3
packaging==19.2
pamqp==2.3.0
pathspec==0.7.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pydantic==1.3
pyflakes==2.1.1
pylint==2.4.4
pymongo==3.10.0
pyparsing==2.4.6
pytest==5.3.2
pytest-asyncio==0.10.0
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
wcwidth==0.1.8
wrapt==1.11.2
yarl==1.4.2
